requirements.txt
requests==2.31.0
python-dotenv==1.0.1
colorama==0.4.6
tabulate==0.9.0
publicsuffix2==2.20191221
dnspython==2.6.1
